Simplify angles and smaller sine table
Use Q0.8 fixed point ratios of Tau to describe a circle, meaning the
full 8 bits. This simplifies adding angles significantly.

Also use Q0.7 for the sine table, this basically encodes "1" as 0x7f
by biasing the number by one. This has the advantage that the table
values are single bytes, it gives what I judge to be good enough
spread of values and it can express negatives values too. The
disadvantage is that it does not have a proper value for "one", but
actually it has, it just do not look for it now. It can potentially
cause some round-off error for straight angles, but I will deal with
such issues later, if it is an issue.